🎨 Split SortButton and SortDropdown in two components and separate dropdowns
