+++ apps/web/src/components/Sort/index.ts
@@ -1,5 +1,5 @@
-export * from './SortByDecadeButton';
-export * from './SortByGenreButton';
-export * from './SortByPopularButton';
-export * from './SortByRatingButton';
-export * from './SortByServiceButton';
+export * from './SortButton';
+export * from './SortDropdown';
+export * from './DecadeDropdown';
+export * from './GenreDropdown';
+export * from './ServiceDropdown';
